feat(后台管理): 商品管理 完善添加商品功能开发

+++ waynboot-admin-api/src/main/java/com/wayn/admin/api/domain/shop/GoodsAttribute.java
@@ -30,7 +30,7 @@ public class GoodsAttribute extends ShopBaseEntity implements Serializable {
     /**
      * 商品表的商品ID
      */
-    private Integer goodsId;
+    private Long goodsId;
 
     /**
      * 商品参数名称
